decode to bg thread + fix crash on iOS
